Upload Source Code
The aim of this project is to build interactively the kinematic chain of
robot from library
of joints and rigid bodies and to output the table of geometric
parameters . This involves
visualization of dierent parameters as well as local body xed frames.
The algorithms for
automatic arrangement of frames must obey modied Denavit Hartenberg
notation rules .
The program must be able to visualize all the structures dened by
SYMORO+ and should
be open-source. This program allows us to compare the software model
with hand written
parameterization thereby reducing the possibilty of modeling errors .